refactor: split multiple React components of TabHome into separate files
