+++ src/pages/TabHome/Calendar/ViewCalendar.tsx
@@ -1,0 +1,125 @@
+import { useContext } from "react";
+import { useTranslation } from "react-i18next";
+import { CyclesContext, ThemeContext } from "../../../state/Context";
+import {
+  getForecastPeriodDates,
+  getOvulationDates,
+  getPeriodDates,
+} from "../../../state/CalculationLogics";
+import {
+  addMonths,
+  endOfMonth,
+  formatISO,
+  max,
+  parseISO,
+  startOfMonth,
+  startOfToday,
+} from "date-fns";
+import { IonButton, IonButtons, IonDatetime } from "@ionic/react";
+import { format } from "../../../utils/datetime";
+import { getCurrentTranslation } from "../../../utils/translation";
+import { CalendarProps } from "./interfaces";
+
+const ViewCalendar = (props: CalendarProps) => {
+  const { t } = useTranslation();
+  const { cycles } = useContext(CyclesContext);
+  const theme = useContext(ThemeContext).theme;
+
+  const periodDates = getPeriodDates(cycles);
+  const forecastPeriodDates = getForecastPeriodDates(cycles);
+  const ovulationDates = getOvulationDates(cycles);
+
+  const firstPeriodDay = periodDates
+    .sort((left, right) => {
+      const leftDate = new Date(left);
+      const rightDate = new Date(right);
+      return leftDate.getTime() - rightDate.getTime();
+    })
+    .at(0);
+
+  const firstPeriodDayDate = firstPeriodDay
+    ? parseISO(firstPeriodDay)
+    : startOfToday();
+
+  const minDate = formatISO(startOfMonth(firstPeriodDayDate));
+
+  const lastForecastPeriodDay = forecastPeriodDates
+    .sort((left, right) => {
+      const leftDate = new Date(left);
+      const rightDate = new Date(right);
+      return leftDate.getTime() - rightDate.getTime();
+    })
+    .at(-1);
+
+  const lastForecastPeriodDayDate = lastForecastPeriodDay
+    ? endOfMonth(parseISO(lastForecastPeriodDay))
+    : endOfMonth(startOfToday());
+
+  const maxDate = formatISO(
+    endOfMonth(max([lastForecastPeriodDayDate, addMonths(startOfToday(), 6)])),
+  );
+
+  return (
+    <IonDatetime
+      className={
+        ovulationDates.includes(format(startOfToday(), "yyyy-MM-dd"))
+          ? `view-calendar-today-ovulation-${theme}`
+          : `view-calendar-${theme}`
+      }
+      presentation="date"
+      locale={getCurrentTranslation()}
+      size="cover"
+      min={minDate}
+      max={maxDate}
+      firstDayOfWeek={1}
+      highlightedDates={(isoDateString) => {
+        if (cycles.length === 0) {
+          return undefined;
+        }
+        if (forecastPeriodDates.includes(isoDateString)) {
+          if (theme === "dark") {
+            return {
+              textColor: `#ffffff`,
+              backgroundColor: `rgba(var(--ion-color-light-${theme}-rgb), 0.3)`,
+            };
+          }
+          return {
+            textColor: `var(--ion-color-dark-${theme})`,
+            backgroundColor: `rgba(var(--ion-color-light-${theme}-rgb), 0.3)`,
+          };
+        } else if (periodDates.includes(isoDateString)) {
+          return theme === "dark"
+            ? {
+                textColor: `#ffffff`,
+                backgroundColor: `rgba(var(--ion-color-dark-${theme}-rgb), 0.6)`,
+              }
+            : {
+                textColor: `var(--ion-color-dark-${theme})`,
+                backgroundColor: `rgba(var(--ion-color-light-${theme}-rgb), 0.8)`,
+              };
+        } else if (ovulationDates.includes(isoDateString)) {
+          return {
+            textColor: `var(--ion-color-ovulation-${theme})`,
+            backgroundColor: `var(--ion-color-calendar-${theme})`,
+            fontWeight: "bold",
+          };
+        }
+
+        return undefined;
+      }}
+    >
+      <IonButtons slot="buttons">
+        <IonButton
+          color={`dark-${theme}`}
+          onClick={() => {
+            props.setIsEditCalendar(true);
+          }}
+        >
+          {t("edit")}
+        </IonButton>
+      </IonButtons>
+    </IonDatetime>
+  );
+};
+
+export default ViewCalendar;
