+++ src/pages/TabHome/Calendar/interfaces.ts
@@ -1,0 +1,3 @@
+export interface CalendarProps {
+  setIsEditCalendar: (newIsOpen: boolean) => void;
+}
