+++ src/pages/TabHome/TabHome.tsx
@@ -1,0 +1,158 @@
+import { useContext, useEffect, useState } from "react";
+import {
+  IonContent,
+  IonPage,
+  IonLabel,
+  useIonRouter,
+  IonButton,
+  IonCol,
+} from "@ionic/react";
+import { App } from "@capacitor/app";
+import { Capacitor } from "@capacitor/core";
+import { useTranslation } from "react-i18next";
+
+import { CyclesContext, ThemeContext } from "../../state/Context";
+import { storage } from "../../data/Storage";
+import { configuration } from "../../data/AppConfiguration";
+
+import Welcome from "../../modals/WelcomeModal";
+import InfoModal from "../../modals/InfoModal";
+
+import {
+  getDaysBeforePeriod,
+  getNewCyclesHistory,
+  getPeriodDatesWithNewElement,
+  isPeriodToday,
+} from "../../state/CalculationLogics";
+import InfoButton from "./InfoButton";
+import ViewCalendar from "./Calendar/ViewCalendar";
+import EditCalendar from "./Calendar/EditCalendar";
+import DemoAlert from "./DemoAlert";
+
+const TabHome = () => {
+  const theme = useContext(ThemeContext).theme;
+
+  const [isInfoModal, setIsInfoModal] = useState(false);
+  const [isWelcomeModal, setIsWelcomeModal] = useState(false);
+  const [isEditCalendar, setIsEditCalendar] = useState(false);
+
+  const router = useIonRouter();
+
+  useEffect(() => {
+    storage.get.cycles().catch((err) => {
+      console.error(`Can't get cycles ${(err as Error).message}`);
+      setIsWelcomeModal(true);
+    });
+  }, []);
+
+  useEffect(() => {
+    const backButtonHandler = () => {
+      if (isInfoModal) {
+        setIsInfoModal(false);
+        router.push("/home");
+        return;
+      }
+      if (!Capacitor.isPluginAvailable("App")) {
+        return;
+      }
+      App.exitApp?.().catch((err) => console.error(err));
+    };
+
+    document.addEventListener("ionBackButton", backButtonHandler);
+
+    return () => {
+      document.removeEventListener("ionBackButton", backButtonHandler);
+    };
+  }, [router, isInfoModal]);
+
+  const { t } = useTranslation();
+  const { cycles, updateCycles } = useContext(CyclesContext);
+
+  return (
+    <IonPage
+      style={{ backgroundColor: `var(--ion-color-background-${theme})` }}
+    >
+      {configuration.features.demoMode && <DemoAlert />}
+      <div
+        id="wide-screen"
+        className={theme}
+      >
+        <IonContent
+          className="ion-padding"
+          color={`transparent-${theme}`}
+        >
+          <Welcome
+            isOpen={isWelcomeModal}
+            setIsOpen={setIsWelcomeModal}
+          />
+          <div id="context-size">
+            <div style={{ marginTop: "30px", marginBottom: "30px" }}>
+              <IonLabel>
+                <p
+                  style={{
+                    fontSize: "35px",
+                    color: `var(--ion-color-text-${theme})`,
+                  }}
+                >
+                  {getDaysBeforePeriod(cycles).title}
+                </p>
+              </IonLabel>
+            </div>
+            <div>
+              <IonLabel>
+                <p
+                  style={
+                    cycles.length === 1
+                      ? {
+                          fontWeight: "bold",
+                          fontSize: "35px",
+                          color: `var(--ion-color-dark-${theme})`,
+                          marginBottom: "30px",
+                        }
+                      : {
+                          fontWeight: "bold",
+                          fontSize: "40px",
+                          color: `var(--ion-color-dark-${theme})`,
+                          marginBottom: "30px",
+                        }
+                  }
+                >
+                  {getDaysBeforePeriod(cycles).days}
+                </p>
+              </IonLabel>
+            </div>
+            <InfoButton setIsInfoModal={setIsInfoModal} />
+            <InfoModal
+              isOpen={isInfoModal}
+              setIsOpen={setIsInfoModal}
+            />
+            <IonCol style={{ marginBottom: "20px" }}>
+              <IonButton
+                className="main"
+                color={`dark-${theme}`}
+                disabled={isPeriodToday(cycles)}
+                onClick={() => {
+                  const newCycles = getNewCyclesHistory(
+                    getPeriodDatesWithNewElement(cycles),
+                  );
+                  updateCycles(newCycles);
+                }}
+              >
+                {t("mark")}
+              </IonButton>
+            </IonCol>
+            <IonCol>
+              {isEditCalendar ? (
+                <EditCalendar setIsEditCalendar={setIsEditCalendar} />
+              ) : (
+                <ViewCalendar setIsEditCalendar={setIsEditCalendar} />
+              )}
+            </IonCol>
+          </div>
+        </IonContent>
+      </div>
+    </IonPage>
+  );
+};
+
+export default TabHome;
