+++ src/pages/TabHome/index.ts
@@ -1,0 +1,3 @@
+import TabHome from "./TabHome";
+
+export default TabHome;
